# Reimplement activity variable builder