--- a/lib/orbf/rules_engine/builders/activity_formula_values_expander.rb
+++ b/lib/orbf/rules_engine/builders/activity_formula_values_expander.rb
@@ -1,14 +1,15 @@
 # frozen_string_literal: true
-
 module Orbf
   module RulesEngine
     class ActivityFormulaValuesExpander
       include VariablesBuilderSupport
 
-      def initialize(package_code, activity_code, formula, orgunit, period)
+      def initialize(package_code, activity_code, expression, values_dependencies, rule_kind, orgunit, period)
         @package_code = package_code
         @activity_code = activity_code
-        @formula = formula
+        @expression = expression
+        @rule_kind = rule_kind
+        @values_dependencies = values_dependencies
         @orgunit = orgunit
         @period = period
       end
@@ -18,24 +19,24 @@ module Orbf
       #    activity_code_code_for_orgunit_id_and_period_2
       # in the expression formula for a given orgunit, period and activity code
       def expand_values
-        expanded_string = formula.expression
+        expanded_string = expression
         spans_subsitutions.each do |k, v|
-          expanded_string = expanded_string.gsub("%{#{k}}", v)
+          expanded_string.gsub!("%{#{k}}", v)
         end
         expanded_string
       end
 
       private
 
-      attr_reader :package_code, :activity_code, :formula, :orgunit, :period
+      attr_reader :package_code, :activity_code, :rule_kind, :expression, :orgunit, :period, :values_dependencies
 
       # return hash with values to susbstitute in the formula
       #   key is dependency symbol and values is the joined values
       #       activity_code_code_for_orgunit_id_and_period_1,
       #       activity_code_code_for_orgunit_id_and_period_2
       def spans_subsitutions
-        formula.values_dependencies.each_with_object({}) do |dependency, hash|
-          span = Orbf::RulesEngine::Spans.matching_span(dependency, formula.rule.kind)
+        values_dependencies.each_with_object({}) do |dependency, hash|
+          span = Orbf::RulesEngine::Spans.matching_span(dependency, rule_kind)
           next unless span
           next if hash[dependency.to_sym]