--- a/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
+++ b/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require_relative "./substitution_builder"
 module Orbf
   module RulesEngine
@@ -10,6 +11,9 @@ module Orbf
         @all_orgunits = orgunits
         @orgunits = orgunits.out_list
         @period = period
+        @tokens = {}
+        @period_facts = Orbf::RulesEngine::PeriodFacts.for(period)
+        @package_rule_codes = package.package_rules.flat_map(&:formulas).map(&:code).to_set
       end
 
       def to_variables
@@ -21,42 +25,14 @@ module Orbf
       attr_reader :package, :orgunits, :period
 
       def activity_formula_variables
-        @tokens ||={}
-        period_facts = Orbf::RulesEngine::PeriodFacts.for(period)
-        package_rule_codes = package.package_rules.flat_map(&:formulas).map(&:code).to_set
         orgunits.each_with_object([]) do |orgunit, array|
           package.all_activities_codes.each do |activity_code|
             activity = package.activities.detect { |candidate| candidate.activity_code == activity_code }
             package.activity_rules.each do |rule|
               rule.formulas.each do |formula|
-                @tokens[formula] ||= Orbf::RulesEngine::Tokenizer.tokenize(formula.expression)
-                subs = {}
-                formula.dependencies.each do |dependency|
-                  activity_state = activity.activity_states.find { |s| s.state == dependency}
-                  if period_facts[dependency]
-                    subs[dependency] = period_facts[dependency]
-                  elsif activity_state&.constant?
-                    subs[dependency]=name_constant(activity.activity_code, activity_state.state, period)
-                  elsif dependency.end_with?("_level_1") || dependency.end_with?("_level_2") || dependency.end_with?("_level_3") || dependency.end_with?("_level_4") || dependency.end_with?("_level_5")
-                    level = dependency[-1].to_i
-                    parent_id = orgunit.parent_ext_ids[level - 1]
-                    subs[dependency]=suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
-                  elsif dependency.end_with?("_level_1_quarterly") || dependency.end_with?("_level_2_quarterly") || dependency.end_with?("_level_3_quarterly") || dependency.end_with?("_level_4_quarterly") || dependency.end_with?("_level_5_quarterly")
-                    level = dependency[-10].to_i
-                    parent_id = orgunit.parent_ext_ids[level - 1]
-                    subs[dependency]=suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
-                  elsif dependency.end_with?("_zone_main_orgunit")
-                    parent_id = @all_orgunits.first.ext_id
-                    subs[dependency]=suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
-                  elsif package_rule_codes.include?(dependency)
-                    subs[dependency]=suffix_for_package(package.code, dependency, orgunit, period)
-                  else
-                    subs[dependency]=suffix_for_id_activity(package.code, activity.activity_code, dependency, orgunit.ext_id, period)
-                  end
-                end
-                instantiated_formula = @tokens[formula].map {|token| subs[token] || token }.join()
-                instantiated_formula = expand_aggregation_values(instantiated_formula, activity)
-                instantiated_formula = Orbf::RulesEngine::ActivityFormulaValuesExpander.new(
+                instantiated_formula = instantiate_formula(formula, activity, orgunit)
+                expand_aggregation_values(instantiated_formula, activity)
+                Orbf::RulesEngine::ActivityFormulaValuesExpander.new(
                   package.code, activity_code,
                   instantiated_formula,
                   formula.values_dependencies,
@@ -69,8 +45,36 @@ module Orbf
         end
       end
 
-      def build_variable(orgunit, activity_code, formula, substitued)
+      def instantiate_formula(formula, activity, orgunit)
+        @tokens[formula] ||= Orbf::RulesEngine::Tokenizer.tokenize(formula.expression)
+        subs = {}
+        formula.dependencies.each do |dependency|
+          activity_state = activity.activity_states.find { |s| s.state == dependency }
+          if @period_facts[dependency]
+            subs[dependency] = @period_facts[dependency]
+          elsif activity_state&.constant?
+            subs[dependency] = name_constant(activity.activity_code, activity_state.state, period)
+          elsif dependency.end_with?("_level_1", "_level_2", "_level_3", "_level_4", "_level_5")
+            level = dependency[-1].to_i
+            parent_id = orgunit.parent_ext_ids[level - 1]
+            subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
+          elsif dependency.end_with?("_level_1_quarterly", "_level_2_quarterly", "_level_3_quarterly", "_level_4_quarterly", "_level_5_quarterly")
+            level = dependency[-10].to_i
+            parent_id = orgunit.parent_ext_ids[level - 1]
+            subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
+          elsif dependency.end_with?("_zone_main_orgunit")
+            parent_id = @all_orgunits.first.ext_id
+            subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
+          elsif @package_rule_codes.include?(dependency)
+            subs[dependency] = suffix_for_package(package.code, dependency, orgunit, period)
+          else
+            subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, orgunit.ext_id, period)
+          end
+        end
+        @tokens[formula].map { |token| subs[token] || token }.join
+      end
 
+      def build_variable(orgunit, activity_code, formula, substitued)
         Orbf::RulesEngine::Variable.new_activity_rule(
           period:                  period,
           key:                     variable_key(package, orgunit, activity_code, formula, period),
@@ -86,12 +90,12 @@ module Orbf
       end
 
       def expand_aggregation_values(instantiated_formula, activity)
-         values_expansions = entities_aggregation_values(activity)
-         values_expansions.each do |k, v|
-           instantiated_formula.gsub!("%{#{k}}", v)
+        values_expansions = entities_aggregation_values(activity)
+        values_expansions.each do |k, v|
+          instantiated_formula.gsub!("%{#{k}}", v)
         end
 
-         instantiated_formula
+        instantiated_formula
       end
 
       def entities_aggregation_values(activity)
@@ -115,7 +119,3 @@ module Orbf
     end
   end
 end
-
-
-
-