--- a/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
+++ b/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
@@ -61,7 +61,7 @@ module Orbf
             parent_id = orgunit.parent_ext_ids[level - 1]
             subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
           elsif dependency.end_with?("_level_1_quarterly", "_level_2_quarterly", "_level_3_quarterly", "_level_4_quarterly", "_level_5_quarterly")
-            level = dependency[-10].to_i
+            level = dependency[-11].to_i
             parent_id = orgunit.parent_ext_ids[level - 1]
             subs[dependency] = suffix_for_id_activity(package.code, activity.activity_code, dependency, parent_id, period)
           elsif dependency.end_with?("_zone_main_orgunit")