--- a/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
+++ b/lib/orbf/rules_engine/builders/activity_formula_variables_builder.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./substitution_builder"
 module Orbf
   module RulesEngine
     class ActivityFormulaVariablesBuilder