before
```
git checkout master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
bundle exec rspec | grep "Finished in "
Finished in 7.47 seconds (files took 0.68991 seconds to load)
bundle exec rspec | grep "Finished in "
Finished in 7.46 seconds (files took 0.6724 seconds to load)
bundle exec rspec | grep "Finished in "
Finished in 7.72 seconds (files took 0.67066 seconds to load)
```

after

```
git checkout reimplement-activity-variable-builder
Switched to branch 'reimplement-activity-variable-builder'
Your branch is up-to-date with 'origin/reimplement-activity-variable-builder'.
bundle exec rspec | grep "Finished in "
Finished in 6.87 seconds (files took 0.70164 seconds to load)
bundle exec rspec | grep "Finished in "
Finished in 6.71 seconds (files took 0.69916 seconds to load)
bundle exec rspec | grep "Finished in "
Finished in 6.68 seconds (files took 0.69875 seconds to load)
```

TODO 
   - [x] add spec for level_x_quarterly 
   - [x] find cause of regression in orbf2 master with this branch (scope activity scope formula with same package formula code)

The performance gain is mainly object allocations reduction. 
On a LMIS like invoice we win 5-7 seconds of 60.
I've running mode on lib_spec to check object allocations with environment variables.
See readme. Before commiting, I was able to compare master and this branch easily.
you can see a drastic drop in string and array allocations.

![image](https://user-images.githubusercontent.com/371692/55242293-38be7780-523d-11e9-9921-fdaffa2fe2eb.png)

this lead me to modify also the variable#dhis2_in_data_element method to reduce string allocations.
the next PR will probably focus on invoice printer and other builders to minimize the tokenizer usage.